+++ docs/source/api/_callbacks.rst
@@ -1,6 +1,9 @@
-=========
+====================================
+Callbacks (includes File Writers)
+====================================
+
 Callbacks
-=========
+------------
 
 .. automodule:: apstools.callbacks.doc_collector
     :members:
@@ -8,3 +11,8 @@ Callbacks
 .. automodule:: apstools.callbacks.snapshot_report
     :members:
 
+
+File Writers
+------------
+
+See the :ref:`filewriters` section.
